move packages based on comments
