+++ internal/controller/promotions/promotions.go
@@ -26,6 +26,7 @@ import (
 	"github.com/akuity/kargo/internal/controller"
 	argocd "github.com/akuity/kargo/internal/controller/argocd/api/v1alpha1"
 	"github.com/akuity/kargo/internal/directives"
+	"github.com/akuity/kargo/internal/event"
 	"github.com/akuity/kargo/internal/indexer"
 	"github.com/akuity/kargo/internal/kargo"
 	"github.com/akuity/kargo/internal/kubeclient"
@@ -393,7 +394,7 @@ func (r *reconciler) Reconcile(
 			msg += fmt.Sprintf(": %s", newStatus.Message)
 		}
 
-		eventAnnotations := libEvent.NewPromotionEventAnnotations(ctx,
+		eventAnnotations := event.NewPromotionAnnotations(ctx,
 			kargoapi.FormatEventControllerActor(r.cfg.Name()),
 			promo, freight)
 
@@ -621,7 +622,7 @@ func (r *reconciler) terminatePromotion(
 		return err
 	}
 
-	eventMeta := libEvent.NewPromotionEventAnnotations(ctx, "", promo, freight)
+	eventMeta := event.NewPromotionAnnotations(ctx, "", promo, freight)
 	eventMeta[kargoapi.AnnotationKeyEventActor] = actor
 
 	r.recorder.AnnotatedEventf(
